+++ spinta/backends/helpers.py
@@ -290,8 +290,10 @@ def flat_select_to_nested(select: Optional[List[str]]) -> SelectTree:
 
 
 def get_model_reserved_props(action: Action) -> List[str]:
-    if action in (Action.GETALL, Action.SEARCH):
+    if action == Action.GETALL:
         return ['_type', '_id', '_revision']
+    elif action == Action.SEARCH:
+        return ['_type', '_id', '_revision', '_base']
     elif action == Action.CHANGES:
         return ['_cid', '_created', '_op', '_id', '_txn', '_revision']
     else: