+++ tests/datasets/test_sql.py
@@ -2367,7 +2367,7 @@ def test_error_handling_server_error(
             'push',
             '-o', 'spinta+https://example.com',
             '--credentials', tmp_path / 'credentials.cfg',
-        ])
+        ], fail=False)
 
     message = (
         'Error when sending and receiving data. Model example/errors/City, '
@@ -2396,7 +2396,7 @@ def test_error_handling_io_error(
             'push',
             '-o', 'spinta+https://example.com',
             '--credentials', tmp_path / 'credentials.cfg',
-        ])
+        ], fail=False)
 
     message = (
         'Error when sending and receiving data. Model example/errors/City, '