+++ tests/test_changes.py
@@ -85,3 +85,41 @@ def test_changes_empty_patch(model, context, app):
     assert send(app, model, ':changes', select=['_cid', '_op', 'status']) == [
         {'_cid': 1, '_op': 'insert', 'status': '1'},
     ]
+
+
+def test_changes_with_ref(context, app):
+    model = 'backends/postgres/country'
+    app.authmodel(model, ['insert'])
+    country = send(app, model, 'insert', {'title': 'Lithuania'})
+
+    model = 'backends/postgres/city'
+    app.authmodel(model, ['insert', 'changes'])
+    send(app, model, 'insert', {'title': 'Vilnius', 'country': {'_id': country.id}})
+    send(app, model, 'insert', {'title': 'Kaunas', 'country': {'_id': country.id}})
+
+    resp = app.get(f"{model}/:changes/:format/ascii")
+    _, header, *lines, _ = resp.text.splitlines()
+    header = header.split()
+    assert header == [
+        '_cid',
+        '_created',
+        '_op',
+        '_id',
+        '_txn',
+        '_revision',
+        'title',
+        'country._id',
+    ]
+    lines = (dict(zip(header, line.split())) for line in lines)
+    lines = [
+        (
+            x['_op'],
+            x['title'],
+            x['country._id'],
+        )
+        for x in lines
+    ]
+    assert lines == [
+        ('insert', 'Vilnius', country.id),
+        ('insert', 'Kaunas', country.id),
+    ]